+++ src/app/components/util/index.js
@@ -1,4 +1,21 @@
-import { ChplEllipsis } from './ellipsis';
+import ChplConfirmation from './confirmation';
+import ChplCriterionTitle from './criterion-title';
+import ChplDialogTitle from './chpl-dialog-title';
+import { ChplEllipsis } from './chpl-ellipsis';
+import ChplHighlightCures from './highlight-cures';
+import ChplLink from './chpl-link';
+import ChplNonProdIndicator from './non-prod-indicator';
 import ChplTextField from './chpl-text-field';
+import ChplTooltip from './chpl-tooltip';
 
-export { ChplEllipsis, ChplTextField };
+export {
+  ChplConfirmation,
+  ChplCriterionTitle,
+  ChplDialogTitle,
+  ChplEllipsis,
+  ChplHighlightCures,
+  ChplLink,
+  ChplNonProdIndicator,
+  ChplTextField,
+  ChplTooltip,
+};