+++ src/app/shared/prop-types/developer.js
@@ -1,9 +1,12 @@
-import { bool, number, oneOfType, shape, string } from 'prop-types';
-import { address, contact } from './';
+import {
+  bool, number, oneOfType, shape, string,
+} from 'prop-types';
+import address from './address';
+import contact from './contact';
 
 const developer = shape({
-  address: address,
-  contact: contact,
+  address,
+  contact,
   developerCode: string,
   developerId: oneOfType([number, string]),
   name: string,
@@ -11,4 +14,4 @@ const developer = shape({
   website: string,
 });
 
-export { developer };
+export default developer;