+++ src/app/shared/prop-types/index.js
@@ -1,6 +1,54 @@
-import { acb } from './acb';
-import { address } from './address';
-import { contact } from './contact';
-import { developer } from './developer';
+import acb from './acb';
+import accessibilityStandard from './accessibility-standard';
+import address from './address';
+import analyticsConfig from './analytics-config';
+import certificationResult from './certification-result';
+import contact from './contact';
+import criterion from './criterion';
+import developer from './developer';
+import qmsStandard from './qms-standard';
+import reliedUponSoftware from './relied-upon-software';
+import resources from './resources';
+import surveillance from './surveillance';
+import surveillanceNonconformity from './surveillance-nonconformity';
+import surveillanceRequirement from './surveillance-requirement';
+import { optionalStandard, selectedOptionalStandard } from './optional-standard';
+import { svap, selectedSvap } from './svap';
+import { testData, selectedTestData } from './test-data';
+import { testFunctionality, selectedTestFunctionality } from './test-functionality';
+import { testProcedure, selectedTestProcedure } from './test-procedure';
+import { testStandard, selectedTestStandard } from './test-standard';
+import { testTool, selectedTestTool } from './test-tool';
+import user from './user';
 
-export { acb, address, contact, developer };
+export {
+  acb,
+  accessibilityStandard,
+  address,
+  analyticsConfig,
+  certificationResult,
+  contact,
+  criterion,
+  developer,
+  qmsStandard,
+  optionalStandard,
+  reliedUponSoftware,
+  resources,
+  selectedOptionalStandard,
+  selectedTestFunctionality,
+  selectedTestProcedure,
+  selectedTestData,
+  selectedTestStandard,
+  selectedTestTool,
+  selectedSvap,
+  surveillance,
+  surveillanceNonconformity,
+  surveillanceRequirement,
+  svap,
+  testFunctionality,
+  testProcedure,
+  testData,
+  testStandard,
+  testTool,
+  user,
+};