+++ src/app/shared/prop-types/surveillance-requirement-type.js
@@ -1,0 +1,9 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import { number, shape, string } from 'prop-types';
+
+const surveillanceRequirementType = shape({
+  id: number,
+  name: string,
+});
+
+export default surveillanceRequirementType;